requirements.txt
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-filter==23.2
django-template-tags==1.1.3
idna==3.4
kombu==5.3.1
oauthlib==3.2.2
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pygismeteo==5.2.0
pygismeteo-base==4.4.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
